Added MonacoForward, a small program to run on ESP32 that can be used as a receiver, and will forward packets in serialized form over a serial communication line to another microcontroller running BBRemotes. The receiver protocol on the other microcontroller then should be MSatProtocol, the newly introduced satellite receiver protocol.
Modified NewDODroid (for now) to use that satellite receiver protocol, so that you can hook up an ESP32 to the droid mainboard's aux serial input and control the droid with ESPnow.

Modified the Mouse code temporarily to do tank drive; this was to support Felix's droid at Comic Con. I think I'll rename that one to "SimpleDroid" at some point, and make it configurable for tank drive (R2) or steered drive (Mouse).

Bunch of other changes.